% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{beta.merMod}
\alias{beta.merMod}
\title{Standardized coeffients of a model.}
\usage{
\method{beta}{merMod}(model, x = TRUE, y = TRUE, skip = NULL)
}
\arguments{
\item{model}{A fitted linear model of type 'merMod' (from the 'lme4' package).}

\item{x}{Logical. Whether or not to standardize predictor variables.}

\item{y}{Logical. Whether or not to standardize criterion variables.}

\item{skip}{A string vector indicating any variables you do \emph{not} wish
to be standarized.}
}
\value{
Returns the summary of a multi-level linear model, with the output
  showing the beta coefficients, standard error, t-values, and p-values for
  each predictor.
}
\description{
\code{beta.merMod} returns the summary of a linear model where all variables
have been standardized.
}
\details{
This function takes a multi-level model and standardizes the variables,
in order to produce standardized (i.e., beta) coefficients rather than
unstandardized (i.e., B) coefficients.

Unlike similar functions, this function properly calculates standardized
estimates for interaction terms (by first standardizing each of the
individual predictor variables).
}
\examples{
# iris data
model <- lmer(Sepal.Width ~ Sepal.Length + Petal.Length + (1|Species), data=iris)
beta(model)  # all variables standardized

beta(model, skip='Petal.Length')  # all variables except Petal.Length standardized
}
\seealso{
\code{\link{beta.lm}}, \code{\link{beta.glm}}, \code{\link{beta.lme}}
}

