% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/block_lm.R
\name{block_lm}
\alias{block_lm}
\title{Incremental block modelling.}
\usage{
block_lm(dv, ..., data = NULL)
}
\arguments{
\item{dv}{The variable name to be used as the dependent variable.}

\item{...}{Pass through variable names (or interaction terms) to add for each
block. To add one term to a block, just pass it through directly; to
add multiple terms, pass it through in a vector or list. Blocks will be
added in the order they are passed to the function, and variables from
previous blocks will be added to each subsequent block.}

\item{data}{An optional data frame containing the variables in the model. If
not found in \code{data}, the variables are taken from the environment from
which the function is called.}
}
\value{
A named list with the following elements:
\tabular{ll}{
  \code{formulas} \tab A list of the regression formulas used for each block.
  \cr
  \code{models} \tab A list of all regression models.\cr
}
}
\description{
\code{block_lm} allows you to incrementally add terms to a linear regression
model.
}
\details{
Given a list of names of variables at each step, this function will run a
series of models, adding the terms for each block incrementally.
}
\examples{
TODO: Need to complete.
}

