% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/block_lm.R
\name{block_lm_q}
\alias{block_lm_q}
\title{Incremental block modelling.}
\usage{
block_lm_q(dv, blocks = NULL, data = NULL)
}
\arguments{
\item{dv}{String of the variable name to be used as the dependent variable.}

\item{blocks}{List of variable names (or interaction terms) to add for each
block. Each list element should be a vector or list of strings with terms
for that block. Variables from previous blocks will be added to each
subsequent block.}

\item{data}{An optional data frame containing the variables in the model. If
not found in \code{data}, the variables are taken from the environment from
which the function is called.}
}
\value{
A named list with the following elements:
\tabular{ll}{
  \code{formulas} \tab A list of the regression formulas used for each block.
  \cr
  \code{models} \tab A list of all regression models.\cr
}
}
\description{
\code{block_lm_q} allows you to incrementally add terms to a linear
regression model.
}
\details{
Given a list of names of variables at each step, this function will run a
series of models, adding the terms for each block incrementally. Note that in
most cases it is easier to use \code{\link{block_lm}} and pass variable names
in directly instead of strings of variable names. \code{block_lm_q} uses
standard evaluation in cases where such evaluation is easier.
}
\examples{
# 2 blocks: Petal.Length; Petal.Length + Petal.Width
model1 <- block_lm_q('Sepal.Length', list('Petal.Length', 'Petal.Width'), data=iris)
summary(model1)
coef(model1)

# 2 blocks: Species; Species + Petal.Length + Petal.Width + Petal.Length:Petal.Width
model2 <- block_lm_q('Sepal.Length', list('Species', 'Petal.Length * Petal.Width')), data=iris)
summary(model2)
coef(model2)
}
\seealso{
\code{\link{block_lm}}
}

