% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cell_means.R
\name{cell_means.aov}
\alias{cell_means.aov}
\title{Estimated values of a linear model.}
\usage{
\method{cell_means}{aov}(model, ..., levels = NULL)
}
\description{
\code{cell_means.aov} is an alias of cell_means.lm.
}
\seealso{
\code{\link{cell_means.lm}}
}

