% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cell_means.R
\name{cell_means}
\alias{cell_means}
\title{Estimated means.}
\usage{
cell_means(model, ...)
}
\arguments{
\item{model}{A fitted linear model of type 'lm'.}

\item{...}{Additional arguments to be passed to the particular method for the
given model.}
}
\value{
The form of the value returned by \code{cell_means} depends on the
  class of its argument. See the documentation of the particular methods for
  details of what is produced by that method.
}
\description{
\code{cell_means} is a generic function for calculating the estimated means
of a linear model.
}
\examples{
TODO: Need to complete.
}
\seealso{
\code{\link{cell_means.lm}}
}

