% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/block_lm.R
\name{fitted.block_lm}
\alias{fitted.block_lm}
\title{Extract model fitted values.}
\usage{
\method{fitted}{block_lm}(model, num = NULL, ...)
}
\arguments{
\item{model}{An object of class "\code{block_lm}", usually, a result of a
call to \code{\link{block_lm}}.}

\item{num}{Numeric vector with the index of model(s) from which to return the
fitted values.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
\code{fitted} method for class "\code{block_lm}".
}
\examples{
TODO: Need to complete.
}
\seealso{
\code{\link{block_lm}}, \code{coef.block_lm},
  \code{residuals.block_lm}
}

