% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graph_model.R
\name{graph_model.aov}
\alias{graph_model.aov}
\title{Graph linear interactions.}
\usage{
\method{graph_model}{aov}(model, y, x, lines = NULL, split = NULL,
  errorbars = "CI", ymin = NULL, ymax = NULL, titles = NULL,
  bargraph = FALSE, draw.legend = TRUE, dodge = 0, exp = FALSE)
}
\description{
\code{graph_model.aov}  is an alias of graph_model.lm.
}
\seealso{
\code{\link{graph_model.lm}}, \code{\link{graph_model.glm}}
}

