% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graph_model.R
\name{graph_model.lm}
\alias{graph_model.lm}
\title{Graph linear interactions.}
\usage{
\method{graph_model}{lm}(model, y, x, lines = NULL, split = NULL,
  errorbars = c("CI", "SE", "none"), ymin = NULL, ymax = NULL,
  labels = NULL, bargraph = FALSE, draw.legend = TRUE, dodge = 0,
  exp = FALSE)
}
\arguments{
\item{model}{A fitted linear model of type 'lm'.}

\item{y}{The variable to be plotted on the y-axis. This variable is required
for the graph.}

\item{x}{The variable to be plotted on the x-axis. This variable is required
for the graph.}

\item{lines}{The variable to be plotted using separate lines (optional).}

\item{split}{The variable to be split among separate graphs (optional).}

\item{errorbars}{A string indicating what kind of error bars to show.
Acceptable values are "CI" (95% confidence intervals), "SE" (+/-1 standard
error of the predicted means), or "none".}

\item{ymin}{Number indicating the minimum value for the y-axis scale. Default
NULL value will adjust position to the lowest y value.}

\item{ymax}{Number indicating the maximum value for the y-axis scale. Default
NULL value will adjust position to the highest y value.}

\item{labels}{A named list with strings for the various plot labels: 'title'
will set the graph title, 'y' sets the y-axis label, 'x' sets the x-axis
label, 'lines' sets the legend label, and 'split' sets the label for the
facet. If any label is not set, the names of the variables will be used.
Setting a label explicitly to NA will set an empty label.}

\item{bargraph}{Logical. TRUE will draw a bar graph of the results; FALSE
will draw a line graph of the results.}

\item{draw.legend}{Logical. Whether or not to draw legend on the graph.}

\item{dodge}{A numeric value indicating the amount each point on the graph
should be shifted left or right, which can help for readability when points
are close together. Default value is 0, with .1 or .2 probably sufficient
in most cases.}

\item{exp}{Logical. If TRUE, the exponential function \code{exp()} will be
used to transform the y-axis (i.e., e to the power of y). Useful for
logistic regressions or for converting log-transformed y-values to their
original units.}
}
\value{
A ggplot2 graph of the plotted variables in the model.
}
\description{
\code{graph_model.lm} provides an easy way to graph interactions in linear
models. Selected variables will be graphed at +/- 1 SD (if continuous) or at
each level of the factor (if categorical).
}
\details{
If there are additional covariates in the model other than what is indicated
to be graphed by the function, these variables will be plotted at their
respective means. In the case of a categorical covariate, the results will be
averaged across all its levels.
}
\examples{
# iris data
model <- lm(Sepal.Width ~ Sepal.Length * Species, data=iris)
graph_model(model, y=Sepal.Width, x=Sepal.Length, lines=Species)
}

