% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graph_model.R
\name{graph_model}
\alias{graph_model}
\title{Graph fitted model interactions.}
\usage{
graph_model(model, ...)
}
\arguments{
\item{model}{A fitted linear model of type 'lm'.}

\item{...}{Additional arguments to be passed to the particular method for the
given model.}
}
\value{
A ggplot2 graph of the plotted variables in the model.
}
\description{
\code{graph_model} provides an easy way to graph interactions in fitted
models. Selected variables will be graphed at +/- 1 SD (if continuous) or at
each level of the factor (if categorical).
}
\examples{
TODO: Need to complete.
}
\seealso{
\code{\link{graph_model.lm}}
}

