% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graph_model.R
\name{graph_model_q.aov}
\alias{graph_model_q.aov}
\title{Graph linear interactions.}
\usage{
graph_model_q.aov(model, y, x, lines = NULL, split = NULL,
  errorbars = "CI", ymin = NULL, ymax = NULL, titles = NULL,
  bargraph = FALSE, draw.legend = TRUE, dodge = 0, exp = FALSE)
}
\description{
\code{graph_model_q.aov}  is an alias of graph_model_q.lm.
}
\seealso{
\code{\link{graph_model_q.lm}}, \code{\link{graph_model_q.glm}}
}

