% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ICC.R
\name{ICC.lme}
\alias{ICC.lme}
\title{Intra-class correlation.}
\usage{
\method{ICC}{lme}(model)
}
\arguments{
\item{model}{A fitted model of type 'lme'.}
}
\value{
The intra-class correlation of the model.
}
\description{
\code{ICC.lme} calculates the intra-class correlation (ICC) from a fitted
multi-level model using the 'nlme' package.
}
\details{
The ICC is the proportion of variance that is between-person variance. For
more information, see
\href{http://davidakenny.net/papers/k&h/MLM_R.pdf}{Hoyt & Kenny (2013)}.
}
\examples{
# iris data
model <- lme(Sepal.Width ~ 1, random=~1|Species, data=iris)
ICC(model)  # .49 of variance is between-subjects
}
\seealso{
\code{\link{ICC.merMod}}
}

