% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ICC.R
\name{ICC.merMod}
\alias{ICC.merMod}
\title{Intra-class correlation.}
\usage{
\method{ICC}{merMod}(model)
}
\arguments{
\item{model}{A fitted model of type 'merMod' (linear, generalized, or
nonlinear).}
}
\value{
The intra-class correlation of the model.
}
\description{
\code{ICC.merMod} calculates the intra-class correlation (ICC) from a fitted
multi-level model using the 'lme4' package.
}
\details{
The ICC is the proportion of variance that is between-person variance. For
more information, see
\href{http://davidakenny.net/papers/k&h/MLM_R.pdf}{Hoyt & Kenny (2013)}.
}
\examples{
TODO: Need to complete.
}

