% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_slopes.R
\name{print.simple_slopes}
\alias{print.simple_slopes}
\title{Print simple slopes.}
\usage{
\method{print}{simple_slopes}(model, digits = max(3L, getOption("digits") -
  3L), signif.stars = getOption("show.signif.stars"), ...)
}
\arguments{
\item{model}{An object of class "\code{simple_slopes}", usually, a result
of a call to \code{\link{simple_slopes}}.}

\item{digits}{The number of significant digits to use when printing.}

\item{signif.stars}{Logical. If \code{TRUE}, 'significance stars' are printed
for each coefficient.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
\code{print} method for class "\code{simple_slopes}".
}
\seealso{
\code{\link{simple_slopes}}
}

