% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_slopes.R
\name{print.simple_slopes_lme4}
\alias{print.simple_slopes_lme4}
\title{Print simple slopes.}
\usage{
\method{print}{simple_slopes_lme4}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class "\code{simple_slopes_lme4}", usually, a result
of a call to \code{\link{simple_slopes}}.}

\item{digits}{The number of significant digits to use when printing.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
\code{print} method for class "\code{simple_slopes_lme4}".
}
\seealso{
\code{\link{simple_slopes}}
}
