% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reghelper.R
\docType{package}
\name{reghelper}
\alias{reghelper}
\alias{reghelper-package}
\title{reghelper: A package to help with running regression analyses.}
\description{
The reghelper package offers numerous functions to make some aspects of
regression models (and similar types of modelling) simpler.
}
\details{
The following methods are currently implemented:
\tabular{ll}{
  \code{\link{beta}} \tab Calculates standardized beta coefficients.\cr
  \code{\link{build_model}} \tab Allows variables to be added to a series of
  regression models sequentially.\cr
  \code{\link{ICC}} \tab Calculates the intra-class correlation for a
  multi-level model.\cr
  \code{\link{cell_means}} \tab Calculate the estimated means for a fitted
  model.\cr
  \code{\link{graph_model}} \tab Easily graph interactions at +/- 1 SD (uses
  ggplot2 package).\cr
  \code{\link{simple_slopes}} \tab Easily calculate the simple effects of an
  interaction.\cr
  \code{\link{sig_regions}} \tab Calculate the Johnson-Neyman regions of
  significance for an interaction.\cr
}
}

