% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/build_model.R
\name{residuals.block_lm_summary}
\alias{residuals.block_lm_summary}
\title{Extract model residuals.}
\usage{
\method{residuals}{block_lm_summary}(model, num = NULL, ...)
}
\arguments{
\item{model}{An object of class "\code{block_lm_summary}", usually, a result
of a call to \code{\link{summary.block_lm}}.}

\item{num}{Numeric vector with the index of model(s) from which to return the
residuals.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The residuals of block(s) `num`, or if `num` is NULL, a list of
  residuals from all blocks.
}
\description{
\code{residuals} method for class "\code{block_lm_summary}".
}
\examples{
# 2 blocks: Petal.Length; Petal.Length + Petal.Width
model1 <- build_model(Sepal.Length, Petal.Length, Petal.Width, data=iris, model='lm')
residuals(summary(model1))  # returns both blocks 1 and 2

# 2 blocks: Species; Species + Petal.Length + Petal.Width + Petal.Length:Petal.Width
model2 <- build_model(Sepal.Length, Species, c(Petal.Length * Petal.Width), data=iris, model='lm')
residuals(summary(model2), num=2)  # returns second block
}
\seealso{
\code{\link{summary.block_lm}}, \code{residuals.block_lm},
  \code{coef.block_lm_summary}
}

