% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_slopes.R
\name{simple_slopes.aov}
\alias{simple_slopes.aov}
\title{Simple slopes of interaction.}
\usage{
\method{simple_slopes}{aov}(model, levels = NULL)
}
\description{
\code{simple_slopes.aov} is an alias of simple_slopes.lm.
}
\seealso{
\code{\link{simple_slopes.lm}}
}

