% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simple_slopes.R
\name{simple_slopes}
\alias{simple_slopes}
\title{Simple slopes of interaction.}
\usage{
simple_slopes(model, ...)
}
\arguments{
\item{model}{A fitted linear model of type 'lm'.}

\item{...}{Additional arguments to be passed to the particular method for the
given model.}
}
\value{
The form of the value returned by \code{simple_slopes} depends on the
  class of its argument. See the documentation of the particular methods for
  details of what is produced by that method.
}
\description{
\code{simple_slopes} is a generic function for calculating the simple effects
of an interaction in a regression model.
}
\examples{
TODO: Need to complete.
}
\seealso{
\code{\link{simple_slopes.lm}}
}

