% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build_model.R
\name{summary.block_glm}
\alias{summary.block_glm}
\title{Summarizing block general linear models.}
\usage{
\method{summary}{block_glm}(model, ...)
}
\arguments{
\item{model}{An object of class "\code{block_glm}", usually, a result of a
call to \code{\link{build_model}}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The function computes and returns a named list of summary statistics
  of the fitted general linear models given in \code{model}. The list has the
  following elements:
\tabular{ll}{
  \code{formulas} \tab A list of the regression formulas used for each block.
  \cr
  \code{family} \tab The link function used in the model fit.\cr
  \code{residuals} \tab A matrix with quantiles of the residuals for each
  model.\cr
  \code{coefficients} \tab A list with a matrix of coefficients for each
  model, as well as the standard error, t-statistic, and p-value.\cr
  \code{overall} \tab A data frame with information about the overall models,
  including the multiple R-squared value; adjusted R-squared; F-statistic,
  degrees of freedom, and p-value for each overall model; and the delta
  R-squared (change in R-squared) and its associated F-statistic and p-value.
  \cr
}
}
\description{
\code{summary} method for class "\code{block_glm}".
}
\examples{
# 2 blocks: Petal.Length; Petal.Length + Petal.Width
model1 <- build_model(Sepal.Length, Petal.Length, Petal.Width, data=iris, model='lm')
summary(model1)
coef(model1)

# 2 blocks: Species; Species + Petal.Length + Petal.Width + Petal.Length:Petal.Width
model2 <- build_model(Sepal.Length, Species, c(Petal.Length * Petal.Width), data=iris, model='lm')
summary(model2)
coef(model2)
}
\seealso{
\code{\link{build_model}}, \code{\link{print.block_glm_summary}}
}

