% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reghelper.R
\name{beta.aov}
\alias{beta.aov}
\title{Standardized coeffients of a model.}
\usage{
\method{beta}{aov}(model, x = TRUE, y = TRUE, skip = NULL)
}
\description{
\code{beta.aov} is an alias of beta.lm.
}
\seealso{
\code{\link{beta.lm}}
}

