% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reghelper.R
\name{beta}
\alias{beta}
\title{Standardized coeffients of a model.}
\usage{
beta(model, ...)
}
\arguments{
\item{model}{A fitted linear model of type 'lm' or 'aov'.}

\item{...}{Additional arguments to be passed to the particular method for the
given model.}
}
\value{
The form of the value returned by \code{beta} depends on the class of
  its argument. See the documentation of the particular methods for details
  of what is produced by that method.
}
\description{
\code{beta} is a generic function for producing standardized coefficients
from regression models.
}
\examples{
TODO: Need to complete.
}
\seealso{
\code{\link{beta.lm}}, \code{\link{beta.aov}}
}

