% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/block_lm.R
\name{residuals.block_lm_summary}
\alias{residuals.block_lm_summary}
\title{Extract model residuals.}
\usage{
\method{residuals}{block_lm_summary}(model, num = NULL, ...)
}
\arguments{
\item{model}{An object of class "\code{block_lm_summary}", usually, a result
of a call to \code{\link{block_lm_summary}}.}

\item{num}{Numeric vector with the index of model(s) from which to return the
residuals.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
\code{residuals} method for class "\code{block_lm_summary}".
}
\examples{
TODO: Need to complete.
}
\seealso{
\code{\link{summary.block_lm}}, \code{residuals.block_lm},
  \code{coef.block_lm_summary}
}

