% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_regions.R
\name{sig_regions.glm}
\alias{sig_regions.glm}
\title{Regions of significance for an interaction.}
\usage{
\method{sig_regions}{glm}(model, alpha = 0.05, precision = 4)
}
\description{
\code{sig_regions.glm} is an alias of sig_regions.lm.
}
\seealso{
\code{\link{sig_regions.lm}}
}

