% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sig_regions.R
\name{sig_regions.lm}
\alias{sig_regions.lm}
\title{Regions of significance for an interaction.}
\usage{
\method{sig_regions}{lm}(model, alpha = 0.05, precision = 4)
}
\arguments{
\item{model}{A fitted linear model of type 'lm' with one two-way interaction
including at least one categorical predictor.}

\item{alpha}{The level at which to test for significance. Default value is
.05.}

\item{precision}{The number of decimal places to which to round the alpha
level (e.g., precision=5 would look for regions of significance at .05000).}
}
\value{
A named vector with a 'lower' and an 'upper' J-N point. If one or
  more of the J-N points fall outside the range of your predictor, the
  function will return NA for that point. If your interaction is not
  significant, both J-N points will be NA.
}
\description{
\code{sig_regions.lm} calculates the Johnson-Neyman (J-N) regions of
significance for an interaction, the points at which the simple effect of the
categorical predictor changes from non-significant to significant.
}
\details{
This function takes a regression model with one two-way interaction, where at
least one of the predictors in the interaction is categorical. If both
variables are categorical, the function will just pick one as the x variable
and one as the categorical moderator. To see the J-N points for the other
categorical predictor, just flip the order of the two variables in your
model.

For more information about regions of significance, see
\href{http://ssrn.com/abstract=2208103}{Spiller, Fitzsimons, Lynch, &
McClelland (2012)}.
}
\examples{
TODO: Need to complete.
}

